auth/requirements.txt
blinker==1.9.0
click==8.2.1
dotenv==0.9.9
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.1
Werkzeug==3.1.3
